Add model properties for authorization and credit
Note property "author" changed to "addedBy", which I used in Quiz and Source models as well.
The "dateAdded" property used now in Note, Quiz, Source, and User models.
User model given "roles" property, which is an array of strings. This will later be validated according to the roles of users in the app.