Fix bug that allowed uniqueId be undefined
In SourceInput and UserInput, uniqueId is a variable to give <input>s a value to append to their id attributes within each render of the component. This is to prevent identical id attributes should the component be rendered multiple times on the same page.
I changed it from just setting on useEffect to setting it with a useState setter.
I also made a function for creating uniqueId values using both the time in milliseconds since Unix Epoch and Math.random(). This should make it incredibly unlikely that any id attribute is identical no matter how many times it is rendered on one page.