Debug MongoDB being called in browser
Still not sure why it happened, but what I did was comment out the lines that imported components from components/index.js, then used instead components/server/index.js or components/client/index.js for importing. This was to keep it clear which components were server and which were client, so that client components would not import server components.

I don't exactly know why, but this solved the errors. My best guess is that when server and client components are exported together, the client components may grab server components accidentally.